requirements.txt
torch
diffusers
transformers
accelerate
matplotlib
Pillow
numpy
pathlib
